add address space & split up into different files
